Fix URL, rename and move mixinsUtils.js, modify a comment
