+++ bower.json
@@ -1,34 +1,34 @@
 {
-  "name": "vue-form-generator",
-  "version": "2.2.2",
-  "homepage": "https://github.com/icebob/vue-form-generator",
-  "authors": [
-    "Icebob"
-  ],
-  "description": "A schema-based form generator component for Vue.js",
-  "main": [
-    "dist/vfg.js",
-    "dist/vfg.css",
-    "dist/vfg-core.js",
-    "dist/vfg-core.css"
-  ],
-  "moduleType": [
-    "amd"
-  ],
-  "keywords": [
-    "vue",
-    "vuejs",
-    "form",
-    "generator",
-    "schema",
-    "json"
-  ],
-  "license": "MIT",
-  "ignore": [
-    "**/.*",
-    "node_modules",
-    "bower_components",
-    "test",
-    "tests"
-  ]
+	"name": "vue-form-generator",
+	"version": "2.2.2",
+	"homepage": "https://github.com/vue-generators/vue-form-generator/",
+	"authors": [
+		"Icebob"
+	],
+	"description": "A schema-based form generator component for Vue.js",
+	"main": [
+		"dist/vfg.js",
+		"dist/vfg.css",
+		"dist/vfg-core.js",
+		"dist/vfg-core.css"
+	],
+	"moduleType": [
+		"amd"
+	],
+	"keywords": [
+		"vue",
+		"vuejs",
+		"form",
+		"generator",
+		"schema",
+		"json"
+	],
+	"license": "MIT",
+	"ignore": [
+		"**/.*",
+		"node_modules",
+		"bower_components",
+		"test",
+		"tests"
+	]
 }
