+++ package.json
@@ -2,7 +2,7 @@
 	"name": "vue-form-generator",
 	"version": "3.0.0-dev",
 	"description": "A schema-based form generator component for Vue.js",
-	"homepage": "https://github.com/icebob/vue-form-generator",
+	"homepage": "https://github.com/vue-generators/vue-form-generator/",
 	"main": "dist/vfg.js",
 	"scripts": {
 		"prebuild": "npm run test",
@@ -38,7 +38,7 @@
 	],
 	"repository": {
 		"type": "git",
-		"url": "https://github.com/icebob/vue-form-generator.git"
+		"url": "https://github.com/vue-generators/vue-form-generator.git"
 	},
 	"files": [
 		"dist/vfg.js",
